Rename stable_random_machine_id to machine_id_v4 / installation_id_v4 (#11138)
## Describe your changes

Rename the new `stableRandomMachineId` to `machineIdV4` to make it clear
(1) we can't rely on `machineIdV3`, and (2) `machineIdV4` is the right
one to use.


[Originally](https://github.com/streamlit/streamlit/commit/b44df19adda5f4418abb95a6dadae30b3ee82a63)
I avoided doing this because the goal is to keep both IDs side by side,
since that will help debug things or even roll back in the future. But
even if we name it `machineIdV4` nothing precludes us from keeping them
side by side.

In the meantime, we're also following bug reports
https://github.com/astral-sh/python-build-standalone/issues/587 and
https://github.com/python/cpython/issues/132710 , which might
rehabilitate `machineIdV3`.

## GitHub Issue Link (if applicable)

n/a

## Testing Plan

- Explanation of why no additional tests are needed: this PR just
renames variables.
- ~~Unit Tests (JS and/or Python)~~
- ~~E2E Tests~~
- ~~Any manual testing needed?~~

---

**Contribution License Agreement**

By submitting this pull request you agree that all contributions to this
project are made under the Apache 2.0 license.