Add server `server_os` and `has_display` to telemetry (#11147)
## Describe your changes

Add server `host_os` and `has_display` to telemetry:
* `server_os`:  "linux" vs "darwin" vs "windows" vs ...
* `has_display`: `true` if the `$DISPLAY` or `$WAYLAND_DISPLAY`
environment variable are set. This is typically -- though not always --
set by the desktop environment (Gnome, KDE, etc). When Streamlit is
running on a server, this is almost guaranteed to be unset.

PS: This also fixes the `server.headless` config option to account for
the somewhat new `$WAYLAND_DISPLAY` environment variable.

PPS: You'll notice that `server_os` in TypeScript is sometimes
`serverOS` and other times `serverOs`. I prefer `serverOS` but
`serverOs` is what protoc prefers. Having both in the code is annoying,
but IMO it's better than unifying on `serverOs`, which just reads like
the name of a cereal: "Server-Ohs!"

## GitHub Issue Link (if applicable)

n/a

## Testing Plan

- Explanation of why no additional tests are needed: the tests would
just be a copy of the implementation
- ~~Unit Tests (JS and/or Python)~~
- ~~E2E Tests~~
- ~~Any manual testing needed?~~

---

**Contribution License Agreement**

By submitting this pull request you agree that all contributions to this
project are made under the Apache 2.0 license.