BugFix 11064: st.dataframe({}) shouldn't show errors (#11100)
## Describe your changes
When an empty dataframe is constructed, the user sees an error message
that says it should never happen. This PR makes an exception for the
case where `numRows` is 0.

## GitHub Issue Link (if applicable)
closes #11064 

## Testing Plan

added e2e tests

---

**Contribution License Agreement**

By submitting this pull request you agree that all contributions to this
project are made under the Apache 2.0 license.