Update the frontend packages to refer to `eslint-plugin-streamlit-custom` via the workspace protocol (#11149)
## Describe your changes

Dependency definitions via the workspace protocol should be more stable
than the file protocol.
For example, if the package is specified via the `file` protocol, the
lock file generated on Windows can be changed from the one from
macOS/Linux and it causes the "The lockfile would have been modified by
this install, which is explicitly forbidden." error.

## GitHub Issue Link (if applicable)

## Testing Plan

- `make mini-devel` succeeded in my local env.
- The current CI pipeline should be sufficient to test this change.

---

**Contribution License Agreement**

By submitting this pull request you agree that all contributions to this
project are made under the Apache 2.0 license.