Activate perf linting rules via ruff (#11185)
## Describe your changes

Activates the [perflint linting rule set from
ruff](https://docs.astral.sh/ruff/rules/#perflint-perf) to lint some
anti-patterns related to performance. The rule set is only activated for
the actual library code and not for tests & internal scripts. The
`PERF203` rule cannot be easily refactored in many of our cases and it
seems somewhat intended in some of these cases that we handle exceptions
within a loop -> added a noqa for those uses.

---

**Contribution License Agreement**

By submitting this pull request you agree that all contributions to this
project are made under the Apache 2.0 license.