Ensure script lifecycle occurs for MPAv1 use case (#11258)
Changes in the MPA v1 handling inadvertently caused issues in clearing
widget state. This is a result of the script lifecycle ending
prematurely. This change forces all pages to run completely and exec in
the case of MPAv0/2

closes #11202

- E2E Tests adds this test case

---

**Contribution License Agreement**

By submitting this pull request you agree that all contributions to this
project are made under the Apache 2.0 license.