Consolidate vertical and horizontal containers (#10960)
## Describe your changes

As part of the Advanced Layouts project, we are extending `st.container`
to be configurable as a flex container. This PR does some refactoring in
preparation for these changes.

Key Changes:

- Create a general `FlexBoxContainer` component that may be vertical or
horizontal. When the flex container changes are rolled out, this will be
configurable by the user, but currently it is just configured in
`layouts.py` to replace the existing vertical container and the
horizontal wrapper for a set of columns.
- Create a general `ContainerContentsWrapper` component that will be
used as a wrapper for the contents of other container types (e.g.
popover, expander) and also in AppView to wrap the contents of the main
app view.

Underneath, both of these components use the same
`StyledFlexContainerBlock` underneath, but the
`ContainerContentsWrapper` is simpler because it doesn't need to handle
the configuration from the user.

These changes should make it easier to reason about the more complex
changes that are going to be added by the advanced layouts project.

This also simplifies the protos required when we rollout the flex
container.


## GitHub Issue Link (if applicable)

## Testing Plan

- Explanation of why no additional tests are needed
- Unit Tests (JS and/or Python)
- E2E Tests
- Any manual testing needed?

---

**Contribution License Agreement**

By submitting this pull request you agree that all contributions to this
project are made under the Apache 2.0 license.