Activate all ruff linter rules and switch to opt-out model (#11350)
## Describe your changes

We have activated almost every ruff rule that is reasonable for our
codebase. This PR changes our ruff config from opt-in - explicitly
select rules - to opt-out - activate all and ignore the ones that we
don't want to support. This enforces active decisions to ignore or
implement new rule set that become available with ruff updates.

---

**Contribution License Agreement**

By submitting this pull request you agree that all contributions to this
project are made under the Apache 2.0 license.