Allow backend URLs to be configurable via window variables (#11378)
## Describe your changes

In some new (Snowflake-internal) deployments of Streamlit, we want to be able to configure an
app to load its `index.html` from one origin but make requests to a Streamlit server at another.
This PR enables this by reading the `window.__STREAMLIT_BACKEND_BASE_URL` and
`window.__STREAMLIT_HOST_CONFIG_BASE_URL` variables and sending requests meant
for the Streamlit server to those URLs if the variables are set. Otherwise, we continue to assume
that the server can be reached at `window.location.href` as usual.

The `index.html` served by this deployment will then be modified to set these variables in
`<script>` tags to configure these URLs.

This may also eventually be useful more generally by the Open Source community in some scenarios
where an app developer, for example, wants to serve the static assets of their app from a CDN while
hosting their Streamlit server at a different origin, but we'll need to make improvements to our
documentation + provide some additional tooling before we can consider this broadly supported.

Note that:

* We also made a small refactor to get rid of the `isHttps` helper function since it's redundant on top
   of the `protocol` field that's already set when parsing a URL.
* When `window.__STREAMLIT_BACKEND_BASE_URL` is set, we currently expect there to be some weirdness
   that happens with multipage apps and the browser's forward/back buttons, but given this behavior isn't
   exposed unless you're changing Streamlit internals, and since fixing this will require some thought, we punt
   on it for now.

## Testing Plan

- Unit Tests (JS and/or Python)
   - Added JS unit tests 
- Any manual testing needed?
- We're not quite at the point where we can test this e2e just yet and
won't be for a few weeks, but merging this code should still be fine
since behavior will be totally unchanged in the regular PyPI
distribution of Streamlit.

---

**Contribution License Agreement**

By submitting this pull request you agree that all contributions to this
project are made under the Apache 2.0 license.