Remove custom constraint setup (#11403)
## Describe your changes

Removes all remaining pieces of the custom constraint setup we have been
using.
Usage of constraints was turned off in
https://github.com/streamlit/streamlit/pull/10512, which makes uploading
those constraint files unnecessary.

If we want to introduce constraint/lock files again, we should consider
migrating the Streamlit repository to UV, which has a locking mechanism.

---

**Contribution License Agreement**

By submitting this pull request you agree that all contributions to this
project are made under the Apache 2.0 license.