Enforce max line length in unit tests (#11460)
## Describe your changes

Enforces the max line length of 120 chars in our unit tests and removes
the E501 rule ignore. There are a couple of files where its not very
practical to enforce this rule, added a module-level ignore in those
files.

---

**Contribution License Agreement**

By submitting this pull request you agree that all contributions to this
project are made under the Apache 2.0 license.