Serve manifest.json from static asset handler (#11462)
## Describe your changes

We need to be able to know the file paths to some of our static assets (in particular
our js and css entrypoints) in our deployments where our `index.html` isn't served
by the Tornado server.

This can be done with nearly no work on our end by tweaking our vite config to generate
a manifest when building the frontend, then serving this file using the existing static asset
handler.

## Testing Plan

- Manual tests
   - Built the package, ran the Streamlit server, and verified that running `curl localhost:8501/manifest.json`
      returns the manifest file as expected.
- Also added a simple test in the PR preview CI stage to check that the file exists and some expected
   keys are in the JSON object

---

**Contribution License Agreement**

By submitting this pull request you agree that all contributions to this
project are made under the Apache 2.0 license.