Tighten up sidebar width (and therefore padding, scrollbars, popover widgets) (#11412)
## Describe your changes

I wanted to make the sidebar a bit narrower, but for that I had to
implement a collection of visual improvements:

- Sidebar
  - Make sidebar narrower
    - **Was** 336px
    - **Now** 256px.
  - Make sidebar padding smaller
- **Was** 1rem + scrollbar gutter -- too large, especially with this
narrower sidebar!!!
- **Now** we want 1rem _already accounting_ for scrollbar, but there are
problems. So I did this:
       - When overlay scrollbars: 1rem
- When regular scrollbars: 2x the scrollbar width (we could easily make
this 1rem total, but [a Chrome
bug](https://issues.chromium.org/issues/40064879) blocks us here)
- Scrollbars
  - **Was**
- On browsers where `::-webkit-scrollbar` is supported, use that to
customize scrollbars.
    - On other browsers, allow the OS to do whatever
- We had no idea which case was which, and couldn't change design
programmatically for each.
    - Besides, `::-webkit-scrollbar` is deprecated
  - **Now**
- Detect whether OS has overlay scrollbars. If yes, just use that (don't
customize scrollbar)
- If no, make scrollbars thinner and prettier with modern CSS, and make
them hide when not hovering.
- For browsers like Safari where you can't change the scrollbar color
using modern CSS, and can't show/hide scrollbars on hover, make
scrollbar thinner but not prettier. And show/hide scrollbar using
CSS-based cropping.
- Dateinput, timeinput, selectbox, multiselect: allow their popups to go
beyond the sidebar boundaries, or they'd be squeezed in side the
sidebar. For this, I just copied what we were already doing with
`st.popover`.

**Why so many screenshots:** the change to the scrollbars modifies the
size of the useful content area, which causes a ton of screenshots to
fail. But I looked through _every single one_ and they seem correct to
me.

## GitHub Issue Link (if applicable)

n/a

## Testing Plan

- Explanation of why no additional tests are needed: these areas are
well-covered by existing tests
- ~~Unit Tests (JS and/or Python)~~
- E2E Tests: screenshots will be updated in this PR
- ~~Any manual testing needed?~~

---

**Contribution License Agreement**

By submitting this pull request you agree that all contributions to this
project are made under the Apache 2.0 license.