[Fix] Enum formatting not working with styled dataframes (#11049)
## Describe your changes

The issue was that after styling a streamlit dataframe with pandas
styler when retrieving the "display_value" of an enum object, the result
returned was the object's name rather than its actual value. This has
now been corrected, and the proper value of the enum object is returned
as expected.

To achieve this, a simple check was added to determine whether the
object is of this specific type, and if so, its actual value is
correctly returned.

## GitHub Issue Link (if applicable)
Fixes #10637 

## Testing Plan

I added a new unit test that reflects the behavior described in the
issue, although in a simplified form.
This test was failing prior to the fix.

---

**Contribution License Agreement**

By submitting this pull request you agree that all contributions to this
project are made under the Apache 2.0 license.

---------

Co-authored-by: Lukas Masuch <Lukas.Masuch@gmail.com>