Fully support `weight` and `unicode-range` in `@font-face` declaration (#11248)
## Describe your changes
Currently, when declaring a font in `[[theme.fontFaces]]` in
`config.toml`, the `weight` field is interpreted as an integer and
`unicode-range` is not supported at all.

This PR supports declaring weight by keyword ("normal", "bold") as well
as declaring weight ranges for variable font files ("100 900").
Additionally, this adds support for unicode range. This may be
beneficial if a font is broken out into different files (e.g. basic and
extended latin characters) and can be helpful for localization. Even if
an app developer doesn't explicitly localize their app, they can define
character sets for other languages that will only conditionally load if
someone uses a browser extension to translate the page, for example.

## GitHub Issue Link (if applicable)
closes #11163 
closes #11247

## Testing Plan
- E2E Tests added for weight and unicode range

---

**Contribution License Agreement**
By submitting this pull request you agree that all contributions to this
project are made under the Apache 2.0 license.

---------

Co-authored-by: Lukas Masuch <103002573+sfc-gh-lmasuch@users.noreply.github.com>
Co-authored-by: lukasmasuch <lukas.masuch@gmail.com>
Co-authored-by: graphite-app[bot] <96075541+graphite-app[bot]@users.noreply.github.com>