[refactor] Utilize useEmotionTheme hook throughout for type-safety (#11573)
## Describe your changes

- Previously, our usages of `useTheme` required an explicit type. This
is a problem since it requires knowledge of common patterns, and wasn't
enforced.
- This PR introduces a single `useEmotionTheme` hook, which handles the
typing in a single place. We also add an eslit rule to prevent any
direct imports of `useTheme` so that it's enforced at lint time.

## GitHub Issue Link (if applicable)

## Testing Plan

- ✅ Lint rule added

---

**Contribution License Agreement**

By submitting this pull request you agree that all contributions to this
project are made under the Apache 2.0 license.