Fix runtime error with event based watcher (#11692)
## Describe your changes

Fix an sporadic error - `RuntimeError: dictionary changed size during
iteration` - that comes up with the event path watcher. Reason is that
it didn't correctly apply the lock mechanism when accessing the watched
paths.

## GitHub Issue Link (if applicable)

- Closes https://github.com/streamlit/streamlit/issues/11691

## Testing Plan

- Added unit tests.

---

**Contribution License Agreement**

By submitting this pull request you agree that all contributions to this
project are made under the Apache 2.0 license.