Add background color to `st.metric` delta (#11678)
## Describe your changes

Slightly changes the design of `st.metric` to add a background color
behind the delta. Colors are the same as for background colored Markdown
text.

Before:
![Screenshot 2025-06-18 at 02 38
37](https://github.com/user-attachments/assets/3c1c06d3-34ff-44bb-8152-96245f758cfb)

After:
![Screenshot 2025-06-20 at 01 28
12](https://github.com/user-attachments/assets/9a90c7c9-754d-4c10-9636-95d54c308832)


## GitHub Issue Link (if applicable)

## Testing Plan

- Updated existing JS unit tests that check for colors to also check for
background colors.
- Updated snapshots. Note that a lot of unrelated snapshots for
`st_columns.py` were updated because `st.metric` is used in the
`st.columns` tests, so the change here pushed the rest of the page down
a bit.

---

**Contribution License Agreement**

By submitting this pull request you agree that all contributions to this
project are made under the Apache 2.0 license.